+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.8.0
+- Fixed issues with delete
+- Fixed sharing transperancy
+- Added a local trash before final delete
+
+# 0.7.0
+- Added upload images/videos & files to blossom servers! 
+- Show images and videos in-line, Files can be downloaded.
+- Fix View-only sharing bug, Where original authors could not edit view-only shared notes.
+
 # 0.5.0
 - Added Tags for pages, allowing users to categorize and organize their notes more effectively.
 - Added relay hints to notes for better visibility.